Add files via upload
In our application the URLRedirector is used to communicate with an application hook. At the time we implemented this widget, timers etc did not work to close the (popup) page that has the widget on it. After the redirect took place, a user had to close the popup manually. In this version I have added the possibility to select a close page microflow. I would like to see this integrated in the main version.